Delinted the workspace: fixed a broken CI clippy/test/build gate and cleared all clippy warnings Touches `.github/workflows/ci.yml`, `crates/app/src/{compose,config_view,other_accounts,reminders,window,assistant,assistant_tools,gtk_test,lookout_view,main,message_list}.rs`, `crates/async-imap/src/mock_stream.rs`, `crates/dav/src/ical.rs`, and `crates/mail/src/{cache,session}.rs`; no data-format change. CI's `clippy`/`test`/`build` jobs all passed `--all-features`, which force-enabled the vendored `async-imap` fork's mutually exclusive `runtime-tokio` and `runtime-async-std` features simultaneously - the crate's own `Cargo.toml` already documents why its default features exclude `runtime-async-std` for exactly this reason - so all three jobs hard-failed to compile (~230 errors) before `-D warnings` ever got a chance to run. The flag is now dropped from all three, matching what `lookout-mail` actually builds with (`runtime-tokio` + `compress`) and what the `greenmail` job already did by scoping its own feature explicitly rather than reaching for `--all-features`. With that gate actually running, clippy reported 15 real warnings across `dav`/`mail`/`app` and one in the vendored `async-imap`: `clippy::cloned_ref_to_slice_refs`, `clippy::needless_borrows_for_generic_args`, `clippy::unnecessary_cast`, and `clippy::io_other_error` were auto-fixed via `cargo clippy --fix`; `sync_mailbox`'s `folders: &mut Vec<Mailbox>` narrowed to `&mut [Mailbox]` (only ever iterated inside the function, never resized); `SecretServiceKeyring::new() -> Arc<dyn KeyringStore>` got an explicit `#[allow(clippy::new_ret_no_self)]` (intentional - a stateless marker type constructed straight into its trait-object form); `UiState`'s two `Rc<dyn Fn(...) -> bool>` callback fields (`follow_up_status`, `follow_up_toggle`) moved behind new `FollowUpStatusFn`/`FollowUpToggleFn` type aliases; and an `ical.rs` doc comment was reflowed so a mid-sentence " - " no longer parsed as a stray Markdown list marker. A repo-wide `cargo fmt --all` (formatting had separately drifted out of sync with `rustfmt` on several files, unrelated to any of the above) brings `cargo fmt --all -- --check` back to clean too.

+++ source/crates/app/src/compose.rs
@@ -1371,7 +1371,7 @@ pub fn build_compose_view(
             // exactly the entries' width. Runs while the popover is
             // mapping, before its size is measured, so the new width lands
             // on the menu that's about to open.
-            let width = (fields_group.width() as i32 / 2).max(240);
+            let width = (fields_group.width() / 2).max(240);
             signature_scroller.set_min_content_width(width);
             (rebuild_handle.borrow())();
         });
